+++ src/components/tree-node/index.ts
@@ -1,0 +1,2 @@
+import TreeNode from './tree-node'
+export default TreeNode
